Change input type from int to float for grades

+++ exercicios/ex007.py
@@ -1,5 +1,5 @@
-n1 = int(input('Digite a primeira nota: '))
-n2 = int(input('Digite a segunda nota: '))
+n1 = float(input('Digite a primeira nota: '))
+n2 = float(input('Digite a segunda nota: '))
 
 media = (n1+n2)/2
